+++ src/main/java/com/coupang/marketplace/global/common/StatusEnum.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.user.controller;
+package com.coupang.marketplace.global.common;
 
 public enum StatusEnum {
